requirements.txt
pandas 
numpy
matplotlib
seaborn
scipy
imbalanced-learn
scikit-learn
optuna
PyYAML==6.0
flask
joblib
evidently
prometheus_client==0.15.0
psutil==5.9.0
dvc
ensure==1.0.2
-e .